put form into 1 component!!!!!!!
